+++ src/NLightning.Domain/Bitcoin/Transactions/Outputs/AnchorOutputInfo.cs
@@ -36,7 +36,7 @@ public class AnchorOutputInfo : IOutputInfo
     /// <summary>
     /// Gets or sets the index of the output in the transaction once it's created.
     /// </summary>
-    public uint? Index { get; set; }
+    public ushort? Index { get; set; }
 
     /// <summary>
     /// Creates a new instance of AnchorOutputInfo.
