+++ src/NLightning.Domain/Bitcoin/Transactions/Outputs/FundingOutputInfo.cs
@@ -1,10 +1,10 @@
-using NLightning.Domain.Bitcoin.Transactions.Enums;
-using NLightning.Domain.Bitcoin.Transactions.Interfaces;
-using NLightning.Domain.Bitcoin.ValueObjects;
-using NLightning.Domain.Crypto.ValueObjects;
-using NLightning.Domain.Money;
-
 namespace NLightning.Domain.Bitcoin.Transactions.Outputs;
+
+using Enums;
+using Interfaces;
+using ValueObjects;
+using Crypto.ValueObjects;
+using Money;
 
 public class FundingOutputInfo : IOutputInfo
 {
@@ -14,7 +14,7 @@ public class FundingOutputInfo : IOutputInfo
 
     public OutputType OutputType => OutputType.Funding;
     public TxId? TransactionId { get; set; }
-    public uint? Index { get; set; }
+    public ushort? Index { get; set; }
 
     public FundingOutputInfo(LightningMoney amount, CompactPubKey localFundingPubKey, CompactPubKey remoteFundingPubKey)
     {
@@ -24,7 +24,7 @@ public class FundingOutputInfo : IOutputInfo
     }
 
     public FundingOutputInfo(LightningMoney amount, CompactPubKey localFundingPubKey, CompactPubKey remoteFundingPubKey,
-                             TxId transactionId, uint index)
+                             TxId transactionId, ushort index)
         : this(amount, localFundingPubKey, remoteFundingPubKey)
     {
         TransactionId = transactionId;
