+++ src/NLightning.Domain/Bitcoin/Transactions/Outputs/ToLocalOutputInfo.cs
@@ -45,7 +45,7 @@ public class ToLocalOutputInfo : IOutputInfo
     /// <summary>
     /// Gets or sets the index of the output in the transaction once it's created.
     /// </summary>
-    public uint? Index { get; set; }
+    public ushort? Index { get; set; }
 
     /// <summary>
     /// Creates a new instance of ToLocalOutputInfo.
