+++ src/NLightning.Domain/Bitcoin/Transactions/Outputs/ToRemoteOutputInfo.cs
@@ -40,7 +40,7 @@ public class ToRemoteOutputInfo : IOutputInfo
     /// <summary>
     /// Gets or sets the index of the output in the transaction once it's created.
     /// </summary>
-    public uint? Index { get; set; }
+    public ushort? Index { get; set; }
 
     /// <summary>
     /// Creates a new instance of ToRemoteOutputInfo.
